move layout to new folder
