+++ client/lib/layout/Footer.tsx
